wip(frontend): Refactor table component
