+++ apps/frontend/src/app/poll/choose-events/choose-events.component.scss
@@ -1,34 +1,3 @@
-.table-responsive td, .table-responsive th {
-  min-width: 100px;
-}
-
-.form-check-hidden {
-  appearance: none;
-  cursor: pointer;
-}
-
 *:not(:hover) > .hover-info {
   visibility: hidden;
-}
-
-@media (min-width: 768px) {
-  th:first-child, th:last-child,
-  td:first-child, td:last-child {
-    position: sticky;
-    backdrop-filter: blur(5px);
-  }
-
-  th:first-child, td:first-child {
-    background: linear-gradient(90deg, var(--bs-body-bg), rgba(var(--bs-body-bg), 0.75));
-    left: 0;
-  }
-
-  th:last-child, td:last-child {
-    background: linear-gradient(90deg, rgba(var(--bs-body-bg), 0.75), var(--bs-body-bg));
-    right: 0;
-  }
-}
-
-h5 {
-  margin: 0;
 }
