+++ apps/frontend/src/app/poll/table/table.component.scss
@@ -1,0 +1,30 @@
+.table-responsive td, .table-responsive th {
+  min-width: 100px;
+}
+
+.form-check-hidden {
+  appearance: none;
+  cursor: pointer;
+}
+
+@media (min-width: 768px) {
+  th:first-child, th:last-child,
+  td:first-child, td:last-child {
+    position: sticky;
+    backdrop-filter: blur(5px);
+  }
+
+  th:first-child, td:first-child {
+    background: linear-gradient(90deg, var(--bs-body-bg), rgba(var(--bs-body-bg), 0.75));
+    left: 0;
+  }
+
+  th:last-child, td:last-child {
+    background: linear-gradient(90deg, rgba(var(--bs-body-bg), 0.75), var(--bs-body-bg));
+    right: 0;
+  }
+}
+
+h5 {
+  margin: 0;
+}
