+++ tsconfig.base.json
@@ -17,7 +17,7 @@
     "paths": {
       "@apollusia/logic": ["libs/types/src/lib/logic/index.ts"],
       "@apollusia/types": ["libs/types/src/index.ts"],
-      "@apollusia/types/*": ["libs/types/src/*"],
+      "@apollusia/types/*": ["libs/types/src/*"]
     }
   },
   "exclude": ["node_modules", "tmp"]
